+++ modules/apps/expando/expando-impl/src/main/java/com/liferay/expando/internal/util/ExpandoBridgeFactoryImpl.java
@@ -14,9 +14,9 @@
 
 package com.liferay.expando.internal.util;
 
+import com.liferay.expando.internal.ExpandoBridgeImpl;
 import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.expando.kernel.util.ExpandoBridgeFactory;
-import com.liferay.portlet.expando.model.impl.ExpandoBridgeImpl;
 
 import org.osgi.service.component.annotations.Component;
